Renamed multi->multipoint
