+++ openquake/hazardlib/source/multi_point.py
